Extract Assets node classes to own file
